remove old BCQ
