+++ compiler/luci/pass/include/luci/CircleOptimizer.h
@@ -33,7 +33,6 @@ public:
     enum Algorithm
     {
       FuseBCQ,
-      FuseNewBCQ,
       FuseInstanceNorm,
       ResolveCustomOpAdd,
       ResolveCustomOpBatchMatMul,
